backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (SQLite - built into Python, SQLAlchemy provides interface)
sqlalchemy==2.0.23
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Astronomical Calculations
pyswisseph==2.10.3.2  # Swiss Ephemeris Python bindings

# Authentication & Security
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
python-dotenv==1.0.0  # Environment variables
bcrypt==4.1.1
cryptography==41.0.7  # Encryption for backups

# HTTP Client (for external APIs like GeoNames)
httpx==0.25.2
aiohttp==3.9.1

# AI / LLM
anthropic==0.40.0  # Claude API for AI-generated interpretations

# Date/Time Handling
python-dateutil==2.8.2
pytz==2023.3.post1

# PDF Generation (for reports)
reportlab==4.0.7
# Alternative: weasyprint==60.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # For async test client

# Development Tools
black==23.12.0  # Code formatter
flake8==6.1.0  # Linter
mypy==1.7.1  # Type checker
isort==5.13.2  # Import sorter

# Production Server
gunicorn==21.2.0  # WSGI HTTP server

# Monitoring & Logging
sentry-sdk==1.38.0  # Error tracking (optional)
python-json-logger==2.0.7  # Structured logging

# Utilities
python-slugify==8.0.1
Pillow==10.1.0  # Image processing (for chart export)

# Backup & Scheduling
APScheduler==3.10.4  # Scheduled backup tasks
keyring>=25.0.0
